pyproject.toml
[project]
name = "behaviour-lock"
version = "0.1.0"
description = "CLI chat agent for legacy code migration"
requires-python = ">=3.11,<3.14"
dependencies = [
    "openai>=1.0.0",
    "chromadb>=0.5.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "prompt_toolkit>=3.0.0",
    "httpx>=0.27.0",
    "pymupdf>=1.24.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[project.scripts]
behaviour-lock = "behaviour_lock.main:cli"
behaviour-lock-api = "behaviour_lock.main:run_api"

[dependency-groups]
dev = [
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/behaviour_lock"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["behaviour_lock"]
